refactor: Extract VentePrivee components from AugmentedModal
Will be simpler to add Ameli AugmentedModal components
